requirements.txt
Pillow==9.0.1
PyYAML==5.4
